Initial commit
This project was created from django-project-templates where the
templates for deployment and frontends were added as feature branches,
which could then be merged to create the starting point for a
djang-based web site. This proved to be unwieldy in practice. Since
there are well defined integration points (the deploy and frontend
directories) it is easier to use submodules, so the branches were
separated out into their own repositories.