Move apps directory up a level to the backend directory
Originally, apps was a sub-directory of project to make search/replace
operations easier, as there was one root directory, project, to search
from, leaving the rest of the backend files and more importantly the
virtualenv directory. However, the recent discovery of "scoped
searches" in PyCharm, rendered this obsolete.

Moving the apps up a level, still keeps the backend directory
uncluttered and now there's the extra bonus that the 'project'
directory is purely for configuration and site-wide features.